Live homepage boxes: leaderboards, recent runs, and stats poll for updates; drop the run counter
