+++ frontend/app/home-revamp.css
@@ -190,9 +190,3 @@
 .rvmp .act-steam:hover { background: #2a475e; }
 .rvmp .act-discord { background: #5865f2; }
 .rvmp .act-discord:hover { background: #4752c4; }
-
-.rvmp .livepulse { display: inline-flex; align-items: center; gap: 7px; font-family: var(--mono); font-size: 11.5px; color: var(--text-3); }
-.rvmp .livepulse .lp-n { color: var(--text); font-weight: 600; }
-.rvmp .lp-dot { width: 7px; height: 7px; border-radius: 50%; background: #2fbf71; animation: lp-pulse 2s ease-out infinite; }
-@keyframes lp-pulse { 0% { box-shadow: 0 0 0 0 rgba(47,191,113,.45); } 70% { box-shadow: 0 0 0 7px rgba(47,191,113,0); } 100% { box-shadow: 0 0 0 0 rgba(47,191,113,0); } }
-@media (prefers-reduced-motion: reduce) { .rvmp .lp-dot { animation: none; } }
